backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
slowapi
redis
langchain
langchain-community
langchain-ollama
chromadb
gitpython
python-dotenv
httpx
